Implement `ConstInstantiable` for arrays and slices
It can be useful to have maps where values are slices or arrays. Given
that Rust stabilized the basic const generics we need on Rust 1.51 and
the project has a MSRV of 1.56, there is no reason I can see for not
implementing this required trait for arrays and slices.

These changes implement such trait for the described types, as long as
the contained type implements `ConstInstantiable` itself.